SES: Breaks up ruletypes into several files
-adds the sonar_uri to each config file
