+++ elastalert/rule_type_definitions/any_rule.py
@@ -1,0 +1,8 @@
+from elastalert.rule_type_definitions.ruletypes import RuleType
+
+class AnyRule(RuleType):
+    """ A rule that will match on any input data """
+
+    def add_data(self, data):
+        for datum in data:
+            self.add_match(datum)
